% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{find_files}
\alias{find_files}
\title{List all files with a certain extension from a URL}
\usage{
find_files(url, extension, url_convert = TRUE)
}
\arguments{
\item{url}{string of URL}

\item{extension}{string of file extension required (without the dot). To get all files set to \code{NULL}.}

\item{url_convert}{boolean. If \code{TRUE} then relative URLs are converted to absolute URLs}
}
\value{
Character vector of URLs
}
\description{
List all files with a certain extension from a URL
}
\examples{
find_files("http://www.abc.net.au/", "htm")
aec_website <- "http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm"
find_files(aec_website, "zip")
\dontrun{
find_files(aec_website, "zip") \%>\% .[c(1,2)] \%>\% download.file(destfile = basename(.))
}
}
