% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{url_convert}
\alias{url_convert}
\title{Convert URLs (relative or absolute) to absolute URLs}
\usage{
url_convert(url, base = attr(url, "base"))
}
\arguments{
\item{url}{string of URL}

\item{base}{string of domain URL}
}
\description{
Convert URLs (relative or absolute) to absolute URLs
}
\examples{
aec_website <- "http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm"
find_files(aec_website, "zip")
find_files("https://catalog.data.gov/dataset?metadata_type=geospatial&_metadata_type_limit=0&q=\%22Congressional\%20Districts\%22", "zip")
}
