% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{download.files}
\alias{download.files}
\title{Download multiple files given a list of URLs}
\usage{
download.files(url, destfile = basename(url))
}
\arguments{
\item{url}{character vector of URLs}

\item{destfile}{character vector of resulting filenames}
}
\value{
VOID
}
\description{
Download multiple files given a list of URLs
}

